wrap matches page component in importable
